Broke up Item into a hierarchy of classes.
