+++ Items.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: b89d6207489f84a87b76567979a9ef89
+folderAsset: yes
+timeCreated: 1497297713
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
