+++ Items/Container.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: feddbaae6d4af4fc28b761876200f5f8
+timeCreated: 1497297713
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
