+++ Items/Decor.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: c5b611b7b340e43fba5f8bb285c6b01c
+timeCreated: 1497298216
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
