+++ Items/Food.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 90775f3121c9b415eba84d81598563c7
+timeCreated: 1497297713
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
