+++ Items/Melee.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: bfc60f42524c444e594fec1625cf28d3
+timeCreated: 1497297713
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
