+++ Items/Projectile.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 3d7050d74ab28429da2416af59f50588
+timeCreated: 1497297713
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
