+++ Items/Ranged.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 5fa56fb188fd0411eb9b9862a0dd485a
+timeCreated: 1497297713
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
