+++ Items/Spawner.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: dab1064ed215e4d95b062177d0e07f97
+timeCreated: 1497297713
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
