+++ Items/WarpDoor.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 7703ffef1ab56483d81fc711719df69e
+timeCreated: 1497297713
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
