+++ Items/Weapon.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: ab3cff364a33d42d79e45fe72af0988e
+timeCreated: 1497297713
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
